refactor: move new ui components into share & update import
