+++ extension/src/prototypes/ui-components/index.ts
@@ -51,7 +51,6 @@ export * from "./InspectorHeader";
 export * from "./InspectorItem";
 export * from "./LayerList";
 export * from "./LayerListItem";
-export * from "./LoadingAnimationIcon";
 export * from "./LoadingScreen";
 export * from "./NumberParameterItem";
 export * from "./OverlayPopper";
@@ -62,7 +61,6 @@ export * from "./ParameterList";
 export * from "./PlateauLogotype";
 export * from "./PlateauSymbol";
 export * from "./PropertyParameterItem";
-export * from "./PropertyActionItem";
 export * from "./QualitativeColorLegend";
 export * from "./QuantitativeColorLegend";
 export * from "./RelaxedPopover";
