+++ extension/src/shared/ui-components/LoadingAnimationIcon.tsx
