+++ extension/src/shared/ui-components/PropertyActionItem.tsx
